test: 테스트 코드 구조 수정
