--- src/test/java/cart/controller/HomeControllerTest.java
@@ -1,42 +1,0 @@
-package cart.controller;
-
-import io.restassured.RestAssured;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.http.HttpStatus;
-
-@AutoConfigureTestDatabase
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class HomeControllerTest {
-
-    @LocalServerPort
-    int port;
-
-    @BeforeEach
-    void setUp() {
-        RestAssured.port = port;
-    }
-
-    @DisplayName("상품 전체 목록을 조회하면 상태코드 200을 반환하는지 확인한다")
-    @Test
-    void getHomeTest() {
-        RestAssured.given().log().all()
-                .when().get("/")
-                .then().log().all()
-                .statusCode(HttpStatus.OK.value());
-    }
-
-    @DisplayName("사용자 전체 목록을 조회하면 상태코드 200을 반환하는지 확인한다")
-    @Test
-    void getSettingTest() {
-        RestAssured.given().log().all()
-                .when().get("/settings")
-                .then().log().all()
-                .statusCode(HttpStatus.OK.value());
-    }
-
-}
