+++ src/test/resources/testData.sql
@@ -1,6 +1,6 @@
+DROP TABLE IF EXISTS cart;
 DROP TABLE IF EXISTS product;
 DROP TABLE IF EXISTS member;
-DROP TABLE IF EXISTS cart;
 
 CREATE TABLE product
 (
